chore: supports build method with custom config
